Refactor dashboard to move widgets into modules. (#5896)
This updates the dashboard in order to move responsibility for dashboard widgets into the modules themselves. It adds the concept of a "widget" to LORIS as a way to provide loosely coupled cross-module dependencies. Modules may now register "widgets" of different types to be used by other modules. As a proof-of-concept, this architecture can be the basis of the upcoming candidates dashboard.

There are 2 types of widgets defined in this PR: dashboard widgets, which add a panel to the dashboard, and usertasks widgets, which add a task to the "My Tasks" panel on the dashboard.

All existing widgets are moved to the most appropriate module that currently exists. The dashboard only maintains ownership of the "Welcome" widget and the "My Tasks" widget itself.
