+++ modules/conflict_resolver/php/module.class.inc
@@ -58,4 +58,40 @@ class Module extends \Module
     {
         return "Conflict Resolver";
     }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @param string $type The type of widgets to get
+     * @param \User  $user The user widgets are being retrieved for.
+     *
+     * @return \LORIS\GUI\Widget[]
+     */
+    public function getWidgets(string $type, \User $user) : array
+    {
+        switch($type) {
+        case 'usertasks':
+            $factory = \NDB_Factory::singleton();
+            $baseURL = $factory->settings()->getBaseURL();
+            $DB      = $factory->database();
+            return [
+                new \LORIS\dashboard\TaskQueryWidget(
+                    $user,
+                    "Data entry conflict",
+                    $DB,
+                    "SELECT COUNT(*) FROM conflicts_unresolved cu
+                         LEFT JOIN flag ON (cu.CommentId1=flag.CommentID)
+                         LEFT JOIN session s ON (flag.SessionID=s.ID)
+                         LEFT JOIN candidate c ON (s.CandID=c.CandID)
+                     WHERE s.Active='Y' AND c.Active='Y' AND
+                         s.CenterID <> 1",
+                    'access_all_profiles',
+                    's.CenterID',
+                    $baseURL . '/' . $this->getName(),
+                    'conflict_resolver'
+                )
+            ];
+        }
+        return [];
+    }
 }
