+++ modules/dashboard/templates/welcomebody.tpl
@@ -1,0 +1,5 @@
+<h3 class="welcome">Welcome, {$username}.</h3>
+<p class="pull-right small login-time">Last login: {$last_login}</p>
+{if !is_null($project_description)}
+<p class="project-description">{$project_description}</p>
+{/if}
