+++ modules/dashboard/templates/welcomefooter.tpl
@@ -1,0 +1,4 @@
+{foreach from=$dashboard_links item=link name=links}
+    <a href="{$link.url}" target="{$link.windowName}">{$link.label}</a>
+    {if !$smarty.foreach.links.last}|{/if}
+{/foreach}
