+++ modules/issue_tracker/php/module.class.inc
@@ -61,4 +61,38 @@ class Module extends \Module
     {
         return "Issue Tracker";
     }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @param string $type The type of widgets to get
+     * @param \User  $user The user widgets are being retrieved for.
+     *
+     * @return \LORIS\GUI\Widget[]
+     */
+    public function getWidgets(string $type, \User $user) : array
+    {
+        switch($type) {
+        case 'usertasks':
+            $factory = \NDB_Factory::singleton();
+            $DB      = $factory->database();
+            $baseURL = $factory->settings()->getBaseURL();
+            return [
+                new \LORIS\dashboard\TaskQueryWidget(
+                    $user,
+                    "Your assigned issue",
+                    $DB,
+                    "SELECT COUNT(*) FROM issues
+                     WHERE status != 'closed' AND assignee="
+                     . $DB->quote($user->getData('UserID')),
+                    '',
+                    '',
+                    $baseURL . "/issue_tracker/?#My%20Issues",
+                    "issue_tracker"
+                )
+            ];
+        }
+        return [];
+
+    }
 }
