+++ modules/mri_violations/php/module.class.inc
@@ -61,4 +61,39 @@ class Module extends \Module
     {
         return "MRI Violated Scans";
     }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @param string $type The type of widgets to get
+     * @param \User  $user The user widgets are being retrieved for.
+     *
+     * @return \LORIS\GUI\Widget[]
+     */
+    public function getWidgets(string $type, \User $user) : array
+    {
+        switch($type) {
+        case "usertasks":
+            $factory = \NDB_Factory::singleton();
+            $DB      = $factory->database();
+            $baseURL = $factory->settings()->getBaseURL();
+
+            return [
+                new \LORIS\dashboard\TaskQueryWidget(
+                    $user,
+                    "Violated scan",
+                    $DB,
+                    "SELECT COUNT(*) FROM mri_protocol_violated_scans
+                        LEFT JOIN candidate c USING (CandID)
+                    WHERE 1 = 1 ",
+                    "violated_scans_view_allsites",
+                    "c.RegistrationCenterID",
+                    $baseURL . "/" . $this->getName(),
+                    ""
+                )
+            ];
+
+        }
+        return [];
+    }
 }
