+++ modules/statistics/css/recruitment.css
@@ -1,0 +1,13 @@
+.progress {
+    height: 40px;
+    margin-bottom: 0;
+}
+
+.progress-bar-male {
+    background-color: #1C70B6;
+}
+
+.progress-bar-female {
+    background-color: #2FA4E7;
+}
+
