+++ modules/user_accounts/php/module.class.inc
@@ -74,4 +74,38 @@ class Module extends \Module
     {
         return "User Accounts";
     }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @param string $type The type of widgets to get
+     * @param \User  $user The user widgets are being retrieved for.
+     *
+     * @return \LORIS\GUI\Widget[]
+     */
+    public function getWidgets(string $type, \User $user) : array
+    {
+        switch($type) {
+        case "usertasks":
+            $factory = \NDB_Factory::singleton();
+            $DB      = $factory->database();
+            $baseURL = $factory->settings()->getBaseURL();
+            return [
+                new \LORIS\dashboard\TaskQueryWidget(
+                    $user,
+                    "Pending account approval",
+                    $DB,
+                    "SELECT COUNT(DISTINCT users.UserID) FROM users
+                        LEFT JOIN user_psc_rel as upr ON (upr.UserID=users.ID)
+                    WHERE users.Pending_approval='Y'
+                        AND (upr.CenterID <> 1 OR upr.CenterID IS NULL)",
+                    'user_accounts_multisite',
+                    'upr.CenterID',
+                    $baseURL . "/user_accounts/?pendingApproval=Y",
+                    "pending-accounts"
+                )
+            ];
+        }
+        return [];
+    }
 }
